backend/requirements.txt
mysqlclient
Django==3.0.2
djangorestframework
django-filter
djangorestframework-simplejwt==4.4.0
django-countries
django-crum
rest_framework_custom_viewsets